CLN: Removing legacy Spark backend
