--- ci/deps/spark.yml
@@ -1,1 +1,0 @@
-pyspark
