+++ docs/source/backends/dask.rst
@@ -1,0 +1,4 @@
+`Dask <https://dask.org/>`_
+===========================
+
+Dask backend is currently experimental.
