+++ setup.cfg
@@ -56,7 +56,6 @@ markers =
     pyspark
     skip_backends
     skip_missing_feature
-    spark
     sqlite
     superuser
     udf
